refactor(backend): split table_row_query + extract table_service schema/migration modules
Two long single-file modules introduced on this branch grew past the team's
own ~1000-line split threshold (cf. the earlier akb-client query-builder
split). Break each into focused modules with a one-directional dependency
DAG; behavior is unchanged (same functions, same call graph).

table_row_query.py (1037 → 145, thin facade):
- row_query_base.py    shared constants/dataclasses/operand/value/select/
                       order/finish primitives (no pipeline deps)
- row_query_string.py  PostgREST querystring compile pipeline
- row_query_ast.py     JSON-AST compile pipeline
- row_query_shape.py   result shaping
Facade keeps select_rows/query_rows and re-exports RowQueryResponse /
compile_row_query / compile_ast_row_query so callers/tests are unchanged.
table_row_write.py now imports the shared helpers from the new modules.

table_service.py (2137 → 1675) A-block extraction:
- table_schema_service.py     merged registry/live schema introspection
- table_migration_service.py  structured idempotent table migrations
Shared helpers (_canonical_pg_type, _fetch_column_meta, index_table_metadata,
alter_table) stay in table_service and are imported by the new modules
(one-directional; no import cycle). Routes and the affected unit tests point
at the new module homes.

The remaining table_service growth (interleaved edits inside the pre-existing
alter_table/create_table) is intentionally left for a separate refactor.

Tests: backend non-e2e suite green (589 passed, 65 skipped); ruff clean.
