+++ backend/app/services/table_row_query.py
@@ -1,89 +1,32 @@
-"""PostgREST-style row-read compiler for vault tables."""
+"""PostgREST-style row-read entry points for vault tables.
+
+Thin orchestration facade: resolve the table, delegate query compilation
+to the querystring pipeline (`row_query_string`) or the JSON-AST pipeline
+(`row_query_ast`), run the compiled SELECT through the user-scoped SQL
+executor, and shape the result (`row_query_shape`). Shared compile
+primitives live in `row_query_base`.
+
+`RowQueryResponse`, `compile_row_query`, and `compile_ast_row_query` are
+re-exported here so existing callers and tests keep importing them from
+`table_row_query`.
+"""
 
 from __future__ import annotations
 
-import json
-import re
 import uuid
-from dataclasses import dataclass
-from datetime import date, datetime
-from decimal import Decimal, InvalidOperation
-from typing import Any, Iterable, Mapping, Sequence
+from typing import Any, Mapping, Sequence
 
 import asyncpg
 
 from app.db.postgres import get_pool
 from app.exceptions import NotFoundError
-from app.repositories import table_data_repo, table_registry_repo
+from app.repositories import table_registry_repo
 from app.services.user_sql_executor import PermissionDeniedError, get_user_sql_executor
-from app.util.errors import (
-    FILTER_TOO_DEEP,
-    INVALID_ARGUMENT,
-    INVALID_CAST,
-    INVALID_FILTER,
-    INVALID_OPERATOR,
-    METHOD_NOT_ALLOWED,
-    NOT_IMPLEMENTED,
-    PERMISSION_DENIED,
-    SQL_ERROR,
-    UNDEFINED_COLUMN,
-    err,
-)
-from app.util.text import fuzzy_hint
-
-
-DEFAULT_LIMIT = 100
-MAX_LIMIT = 1000
-MAX_BOOL_DEPTH = 3
-BOOKKEEPING_COLUMNS = {
-    "id": "uuid",
-    "created_by": "text",
-    "created_at": "timestamp",
-    "updated_at": "timestamp",
-}
-CAST_SQL = {
-    "int": "integer",
-    "numeric": "numeric",
-    "float": "double precision",
-    "bool": "boolean",
-    "date": "date",
-    "timestamp": "timestamp",
-    "uuid": "uuid",
-    "text": "text",
-}
-
-
-def _is_json_type(type_name: str) -> bool:
-    return type_name in {"json", "jsonb"}
-_JSON_PATH_RE = re.compile(
-    r"^(?P<base>[a-z][a-z0-9_]*)(?:(?P<arrow>->>|#>>)(?P<path>[^:]+))?(?:::(?P<cast>[a-z]+))?$"
-)
-
-
-@dataclass
-class RowQueryResponse:
-    body: dict[str, Any]
-    content_range: str | None = None
-
-
-@dataclass
-class _Operand:
-    sql: str
-    params: list[Any]
-    type_name: str
-
-
-@dataclass
-class _Projection:
-    sql: str
-    output_key: str
-    result_key: str
-
-
-@dataclass
-class _Page:
-    limit: int
-    offset: int
+from app.util.errors import PERMISSION_DENIED, SQL_ERROR, err
+from app.services.row_query_ast import compile_ast_row_query
+from app.services.row_query_base import RowQueryResponse
+from app.services.row_query_shape import _shape_result
+from app.services.row_query_string import compile_row_query
 
 
 async def select_rows(
@@ -200,837 +143,3 @@ async def query_rows(
         offset=compiled["page"].offset,
     )
     return RowQueryResponse(body=body, content_range=content_range)
-
-
-def compile_row_query(
-    *,
-    vault_name: str,
-    table_name: str,
-    columns: list[dict],
-    query_params: Sequence[tuple[str, str]],
-    range_header: str | None = None,
-    prefer_header: str | None = None,
-) -> dict[str, Any]:
-    column_meta = _column_meta(columns)
-    params: list[Any] = []
-
-    page_or_error = _parse_page(query_params, range_header)
-    if isinstance(page_or_error, dict):
-        return page_or_error
-    page = page_or_error
-
-    projections_or_error = _compile_select(_last_value(query_params, "select"), column_meta, params)
-    if isinstance(projections_or_error, dict):
-        return projections_or_error
-    projections = projections_or_error
-
-    where_or_error = _compile_filters(query_params, column_meta, params)
-    if isinstance(where_or_error, dict):
-        return where_or_error
-    order_or_error = _compile_order(_last_value(query_params, "order"), column_meta, params)
-    if isinstance(order_or_error, dict):
-        return order_or_error
-
-    return _finish_row_query(
-        vault_name=vault_name,
-        table_name=table_name,
-        projections=projections,
-        where_sql=where_or_error,
-        order_sql=order_or_error,
-        page=page,
-        params=params,
-        count_exact=_prefer_count_exact(prefer_header),
-    )
-
-
-def compile_ast_row_query(
-    *,
-    vault_name: str,
-    table_name: str,
-    columns: list[dict],
-    ast: Mapping[str, Any],
-    range_header: str | None = None,
-    prefer_header: str | None = None,
-) -> dict[str, Any]:
-    forbidden_writes = sorted({"insert", "update", "delete", "upsert"} & set(ast))
-    if forbidden_writes:
-        return err(
-            "POST /query supports read ASTs only.",
-            code=METHOD_NOT_ALLOWED,
-            unsupported_keys=forbidden_writes,
-        )
-
-    column_meta = _column_meta(columns)
-    params: list[Any] = []
-
-    page_or_error = _parse_page(_ast_page_params(ast), range_header)
-    if isinstance(page_or_error, dict):
-        return page_or_error
-    page = page_or_error
-
-    select_or_error = _ast_select_value(ast.get("select"))
-    if isinstance(select_or_error, dict):
-        return select_or_error
-    projections_or_error = _compile_select(select_or_error, column_meta, params)
-    if isinstance(projections_or_error, dict):
-        return projections_or_error
-    projections = projections_or_error
-
-    filter_node = _ast_filter_root(ast)
-    where_sql = ""
-    if filter_node is not None:
-        compiled_filter = _compile_ast_filter(filter_node, column_meta, params, depth=1)
-        if isinstance(compiled_filter, dict):
-            return compiled_filter
-        where_sql = compiled_filter
-
-    order_or_error = _compile_ast_order(ast.get("order"), column_meta, params)
-    if isinstance(order_or_error, dict):
-        return order_or_error
-
-    return _finish_row_query(
-        vault_name=vault_name,
-        table_name=table_name,
-        projections=projections,
-        where_sql=where_sql,
-        order_sql=order_or_error,
-        page=page,
-        params=params,
-        count_exact=_prefer_count_exact(prefer_header) or _ast_count_exact(ast.get("count")),
-    )
-
-
-def _finish_row_query(
-    *,
-    vault_name: str,
-    table_name: str,
-    projections: list[_Projection],
-    where_sql: str,
-    order_sql: str,
-    page: _Page,
-    params: list[Any],
-    count_exact: bool,
-) -> dict[str, Any]:
-    select_sql = ", ".join(p.sql for p in projections)
-    from_sql = f"FROM {table_data_repo.pg_table_name(vault_name, table_name)}"
-    if where_sql:
-        from_sql += f" WHERE {where_sql}"
-
-    if count_exact:
-        page_sql = f"SELECT {select_sql}, TRUE AS __akb_present {from_sql}"
-        if order_sql:
-            page_sql += f" ORDER BY {order_sql}"
-        page_sql += f" LIMIT {page.limit} OFFSET {page.offset}"
-        sql = (
-            f"WITH __akb_count AS (SELECT count(*) AS __akb_total {from_sql}), "
-            f"__akb_page AS ({page_sql}) "
-            "SELECT __akb_page.*, __akb_count.__akb_total "
-            "FROM __akb_count LEFT JOIN __akb_page ON TRUE"
-        )
-    else:
-        sql = f"SELECT {select_sql} {from_sql}"
-        if order_sql:
-            sql += f" ORDER BY {order_sql}"
-        sql += f" LIMIT {page.limit} OFFSET {page.offset}"
-
-    return {
-        "sql": sql,
-        "params": params,
-        "projections": projections,
-        "count_exact": count_exact,
-        "page": page,
-    }
-
-
-def _ast_page_params(ast: Mapping[str, Any]) -> list[tuple[str, str]]:
-    params: list[tuple[str, str]] = []
-    for key in ("limit", "offset"):
-        if key in ast and ast[key] is not None:
-            params.append((key, str(ast[key])))
-    page = ast.get("page")
-    if isinstance(page, Mapping):
-        for key in ("limit", "offset"):
-            if key in page and page[key] is not None:
-                params.append((key, str(page[key])))
-    return params
-
-
-def _ast_select_value(value: Any) -> str | None | dict[str, Any]:
-    if value is None:
-        return None
-    if isinstance(value, str):
-        return value
-    if isinstance(value, Sequence) and not isinstance(value, (str, bytes, bytearray)):
-        out: list[str] = []
-        for item in value:
-            if not isinstance(item, str):
-                return err("AST select entries must be strings.", code=INVALID_ARGUMENT)
-            out.append(item)
-        return ",".join(out)
-    return err("AST select must be a string or string array.", code=INVALID_ARGUMENT)
-
-
-def _ast_filter_root(ast: Mapping[str, Any]) -> Any | None:
-    for key in ("filter", "where"):
-        if key in ast:
-            return ast[key]
-    if any(key in ast for key in ("and", "or", "col", "jsonb")):
-        return ast
-    return None
-
-
-def _compile_ast_filter(
-    node: Any,
-    column_meta: dict[str, str],
-    params: list[Any],
-    *,
-    depth: int,
-) -> str | dict[str, Any]:
-    if not isinstance(node, Mapping):
-        return err("AST filter nodes must be objects.", code=INVALID_ARGUMENT)
-    group_keys = [key for key in ("and", "or") if key in node]
-    if group_keys:
-        if len(group_keys) != 1:
-            return err("AST filter node must contain only one boolean group.", code=INVALID_ARGUMENT)
-        return _compile_ast_bool_group(group_keys[0], node[group_keys[0]], column_meta, params, depth=depth)
-    return _compile_ast_condition(node, column_meta, params)
-
-
-def _compile_ast_bool_group(
-    joiner: str,
-    value: Any,
-    column_meta: dict[str, str],
-    params: list[Any],
-    *,
-    depth: int,
-) -> str | dict[str, Any]:
-    if depth > MAX_BOOL_DEPTH:
-        return err("Boolean filter nesting is too deep.", code=FILTER_TOO_DEEP)
-    if not isinstance(value, Sequence) or isinstance(value, (str, bytes, bytearray)):
-        return err(f"AST {joiner} filter must be an array.", code=INVALID_ARGUMENT)
-    clauses: list[str] = []
-    for item in value:
-        clause_or_error = _compile_ast_filter(item, column_meta, params, depth=depth + 1)
-        if isinstance(clause_or_error, dict):
-            return clause_or_error
-        if clause_or_error:
-            clauses.append(clause_or_error)
-    if not clauses:
-        return err(f"Invalid boolean filter: empty {joiner} group.", code=INVALID_FILTER)
-    glue = " OR " if joiner == "or" else " AND "
-    return f"({glue.join(f'({c})' for c in clauses)})"
-
-
-def _compile_ast_condition(
-    node: Mapping[str, Any],
-    column_meta: dict[str, str],
-    params: list[Any],
-) -> str | dict[str, Any]:
-    operator = node.get("op")
-    if not isinstance(operator, str) or not operator:
-        return err("AST condition must include an operator string.", code=INVALID_ARGUMENT)
-    operand_or_error = _compile_ast_operand(node, column_meta)
-    if isinstance(operand_or_error, dict):
-        return operand_or_error
-    operand = _bind_operand_params(operand_or_error, params)
-    return _compile_ast_operator(operand, operator.lower(), node.get("val"), params)
-
-
-def _compile_ast_operand(node: Mapping[str, Any], column_meta: dict[str, str]) -> _Operand | dict[str, Any]:
-    col = node.get("col")
-    if isinstance(col, str):
-        return _compile_operand(col, column_meta)
-    jsonb = node.get("jsonb")
-    if not isinstance(jsonb, Mapping):
-        return err("AST condition must include col or jsonb.", code=INVALID_ARGUMENT)
-    base = jsonb.get("col")
-    if not isinstance(base, str) or not base:
-        return err("AST jsonb operand must include a column name.", code=INVALID_ARGUMENT)
-    if base not in column_meta:
-        return _unknown_column(base, column_meta)
-    if not _is_json_type(column_meta[base]):
-        return err(f"Column {base!r} is not a JSON column.", code=UNDEFINED_COLUMN)
-    path = jsonb.get("path")
-    if not isinstance(path, Sequence) or isinstance(path, (str, bytes, bytearray)) or not path:
-        return err("AST jsonb operand must include a non-empty path array.", code=INVALID_ARGUMENT)
-    path_items: list[str] = []
-    for item in path:
-        if not isinstance(item, str) or not item:
-            return err("AST jsonb path entries must be non-empty strings.", code=INVALID_ARGUMENT)
-        path_items.append(item)
-    cast = jsonb.get("cast")
-    if cast is not None and (not isinstance(cast, str) or cast not in CAST_SQL):
-        return err(f"Invalid JSON cast {cast!r}.", code=INVALID_CAST, allowed_casts=sorted(CAST_SQL))
-    sql_base = table_data_repo.safe_ident(base)
-    if len(path_items) == 1:
-        expr = f"{sql_base} ->> ${{param}}::text"
-        operand_params: list[Any] = [path_items[0]]
-    else:
-        expr = f"{sql_base} #>> ${{param}}::text[]"
-        operand_params = [path_items]
-    type_name = cast or "text"
-    if cast:
-        expr = f"({expr})::{CAST_SQL[cast]}"
-    return _Operand(sql=expr, params=operand_params, type_name=type_name)
-
-
-def _compile_ast_operator(
-    operand: _Operand,
-    operator: str,
-    value: Any,
-    params: list[Any],
-) -> str | dict[str, Any]:
-    if operator.startswith("not."):
-        inner = _compile_ast_operator(operand, operator.removeprefix("not."), value, params)
-        if isinstance(inner, dict):
-            return inner
-        return f"NOT ({inner})"
-    if operator == "not":
-        if not isinstance(value, Mapping):
-            return err("AST not operator expects an object value.", code=INVALID_ARGUMENT)
-        inner_op = value.get("op")
-        if not isinstance(inner_op, str) or not inner_op:
-            return err("AST not operator expects an inner operator.", code=INVALID_ARGUMENT)
-        inner = _compile_ast_operator(operand, inner_op.lower(), value.get("val"), params)
-        if isinstance(inner, dict):
-            return inner
-        return f"NOT ({inner})"
-    if operator == "is":
-        lowered = "null" if value is None else str(value).lower()
-        if lowered == "null":
-            return f"{operand.sql} IS NULL"
-        if lowered in {"true", "false"}:
-            return f"{operand.sql} IS {lowered.upper()}"
-        return err("is operator only supports null, true, or false.", code=INVALID_FILTER)
-    if operator in {"eq", "neq", "gt", "gte", "lt", "lte", "like", "ilike"}:
-        sql_op = {
-            "eq": "=",
-            "neq": "<>",
-            "gt": ">",
-            "gte": ">=",
-            "lt": "<",
-            "lte": "<=",
-            "like": "LIKE",
-            "ilike": "ILIKE",
-        }[operator]
-        type_name = "text" if operator in {"like", "ilike"} else operand.type_name
-        raw_value = str(value).replace("*", "%") if operator in {"like", "ilike"} else value
-        converted = _convert_ast_value(raw_value, type_name)
-        if isinstance(converted, dict):
-            return converted
-        return f"{operand.sql} {sql_op} {_add_param(params, converted)}"
-    if operator == "in":
-        if not isinstance(value, Sequence) or isinstance(value, (str, bytes, bytearray)):
-            return err("AST in operator expects an array value.", code=INVALID_ARGUMENT)
-        values: list[Any] = []
-        for item in value:
-            converted = _convert_ast_value(item, operand.type_name)
-            if isinstance(converted, dict):
-                return converted
-            values.append(converted)
-        return f"{operand.sql} = ANY({_add_param(params, values)})"
-    if operator == "cs":
-        if not _is_json_type(operand.type_name):
-            return err("cs operator is only supported for JSON columns in this table API.", code=INVALID_OPERATOR)
-        contains_value = json.dumps(value) if isinstance(value, (Mapping, list)) else _parse_contains_value(str(value), "json")
-        if isinstance(contains_value, dict):
-            return contains_value
-        return f"{operand.sql} @> {_add_param(params, contains_value)}::jsonb"
-    return err(f"Unknown row-read operator: {operator}", code=INVALID_OPERATOR)
-
-
-def _compile_ast_order(
-    value: Any,
-    column_meta: dict[str, str],
-    params: list[Any],
-) -> str | dict[str, Any]:
-    if value is None:
-        return ""
-    values = value if isinstance(value, Sequence) and not isinstance(value, (str, bytes, bytearray, Mapping)) else [value]
-    parts: list[str] = []
-    for item in values:
-        if isinstance(item, str):
-            order_or_error = _compile_order(item, column_meta, params)
-            if isinstance(order_or_error, dict):
-                return order_or_error
-            if order_or_error:
-                parts.append(order_or_error)
-            continue
-        if not isinstance(item, Mapping):
-            return err("AST order entries must be strings or objects.", code=INVALID_ARGUMENT)
-        direction = str(item.get("dir") or item.get("direction") or "asc").lower()
-        if direction not in {"asc", "desc"}:
-            return err("AST order direction must be asc or desc.", code=INVALID_ARGUMENT)
-        operand_or_error = _compile_ast_operand(item, column_meta)
-        if isinstance(operand_or_error, dict):
-            return operand_or_error
-        operand = _bind_operand_params(operand_or_error, params)
-        parts.append(f"{operand.sql} {direction.upper()}")
-    return ", ".join(parts)
-
-
-def _ast_count_exact(value: Any) -> bool:
-    if isinstance(value, bool):
-        return value
-    return isinstance(value, str) and value.lower() == "exact"
-
-
-def _convert_ast_value(value: Any, type_name: str) -> Any | dict[str, Any]:
-    if value is None:
-        return err("Filter value cannot be null; use the is operator.", code=INVALID_ARGUMENT)
-    if _is_json_type(type_name) and not isinstance(value, str):
-        return json.dumps(value)
-    return _convert_value(str(value), type_name)
-
-
-def _column_meta(columns: list[dict]) -> dict[str, str]:
-    meta = dict(BOOKKEEPING_COLUMNS)
-    for col in columns:
-        if not isinstance(col, dict):
-            continue
-        name = col.get("name")
-        if isinstance(name, str) and name:
-            meta[name] = str(col.get("type") or "text").lower()
-    return meta
-
-
-def _compile_select(
-    select_value: str | None,
-    column_meta: dict[str, str],
-    params: list[Any],
-) -> list[_Projection] | dict[str, Any]:
-    if not select_value:
-        return [_Projection(sql="*", output_key="*", result_key="*")]
-    projections: list[_Projection] = []
-    for idx, token in enumerate(_split_top_level(select_value)):
-        token = token.strip()
-        if not token:
-            continue
-        if re.search(r"(?<!:):(?!:)", token):
-            return err("Column aliases in select= are not implemented yet.", code=NOT_IMPLEMENTED)
-        if token == "*":
-            projections.append(_Projection(sql="*", output_key="*", result_key="*"))
-            continue
-        operand_or_error = _compile_operand(token, column_meta)
-        if isinstance(operand_or_error, dict):
-            return operand_or_error
-        operand = _bind_operand_params(operand_or_error, params)
-        if operand.sql == token:
-            projections.append(_Projection(sql=operand.sql, output_key=token, result_key=token))
-        else:
-            result_key = f"__akb_col_{idx}"
-            projections.append(_Projection(sql=f"{operand.sql} AS {result_key}", output_key=token, result_key=result_key))
-    return projections or [_Projection(sql="*", output_key="*", result_key="*")]
-
-
-def _compile_filters(
-    query_params: Sequence[tuple[str, str]],
-    column_meta: dict[str, str],
-    params: list[Any],
-) -> str | dict[str, Any]:
-    clauses: list[str] = []
-    for key, value in query_params:
-        if key in {"select", "order", "limit", "offset"}:
-            continue
-        if key in {"or", "and"}:
-            clause_or_error = _compile_bool_group(key, value, column_meta, params, depth=1)
-        else:
-            clause_or_error = _compile_condition(key, value, column_meta, params)
-        if isinstance(clause_or_error, dict):
-            return clause_or_error
-        if clause_or_error:
-            clauses.append(clause_or_error)
-    return " AND ".join(f"({c})" for c in clauses)
-
-
-def _compile_bool_group(
-    joiner: str,
-    value: str,
-    column_meta: dict[str, str],
-    params: list[Any],
-    *,
-    depth: int,
-) -> str | dict[str, Any]:
-    if depth > MAX_BOOL_DEPTH:
-        return err("Boolean filter nesting is too deep.", code=FILTER_TOO_DEEP)
-    inner = value.strip()
-    if inner.startswith("(") and inner.endswith(")"):
-        inner = inner[1:-1]
-    clauses: list[str] = []
-    for part in _split_top_level(inner):
-        part = part.strip()
-        if not part:
-            continue
-        nested = _nested_group(part)
-        if nested:
-            nested_joiner, nested_value = nested
-            clause_or_error = _compile_bool_group(
-                nested_joiner, nested_value, column_meta, params, depth=depth + 1,
-            )
-        else:
-            field, op_value = _split_bool_condition(part)
-            if field is None or op_value is None:
-                return err(f"Invalid boolean filter: {part}", code=INVALID_FILTER)
-            clause_or_error = _compile_condition(field, op_value, column_meta, params)
-        if isinstance(clause_or_error, dict):
-            return clause_or_error
-        clauses.append(clause_or_error)
-    if not clauses:
-        return err(f"Invalid boolean filter: empty {joiner} group.", code=INVALID_FILTER)
-    glue = " OR " if joiner == "or" else " AND "
-    return glue.join(f"({c})" for c in clauses)
-
-
-def _compile_condition(
-    field: str,
-    raw_value: str,
-    column_meta: dict[str, str],
-    params: list[Any],
-) -> str | dict[str, Any]:
-    operand_or_error = _compile_operand(field, column_meta)
-    if isinstance(operand_or_error, dict):
-        return operand_or_error
-    operand = _bind_operand_params(operand_or_error, params)
-    operator, value = _split_operator(raw_value)
-    if operator is None:
-        return err(f"Invalid filter value for {field}: expected op.value", code=INVALID_FILTER)
-    return _compile_operator(operand, operator, value, params)
-
-
-def _compile_operator(
-    operand: _Operand,
-    operator: str,
-    value: str,
-    params: list[Any],
-) -> str | dict[str, Any]:
-    if operator == "not":
-        inner_op, inner_value = _split_operator(value)
-        if inner_op is None:
-            return err("Invalid not filter: expected not.op.value", code=INVALID_FILTER)
-        inner = _compile_operator(operand, inner_op, inner_value, params)
-        if isinstance(inner, dict):
-            return inner
-        return f"NOT ({inner})"
-    if operator == "is":
-        lowered = value.lower()
-        if lowered == "null":
-            return f"{operand.sql} IS NULL"
-        if lowered in {"true", "false"}:
-            return f"{operand.sql} IS {lowered.upper()}"
-        return err("is operator only supports null, true, or false.", code=INVALID_FILTER)
-    if operator in {"eq", "neq", "gt", "gte", "lt", "lte", "like", "ilike"}:
-        sql_op = {
-            "eq": "=",
-            "neq": "<>",
-            "gt": ">",
-            "gte": ">=",
-            "lt": "<",
-            "lte": "<=",
-            "like": "LIKE",
-            "ilike": "ILIKE",
-        }[operator]
-        type_name = "text" if operator in {"like", "ilike"} else operand.type_name
-        converted = _convert_value(value.replace("*", "%") if operator in {"like", "ilike"} else value, type_name)
-        if isinstance(converted, dict):
-            return converted
-        return f"{operand.sql} {sql_op} {_add_param(params, converted)}"
-    if operator == "in":
-        values = _parse_in_values(value, operand.type_name)
-        if isinstance(values, dict):
-            return values
-        return f"{operand.sql} = ANY({_add_param(params, values)})"
-    if operator == "cs":
-        contains_value = _parse_contains_value(value, operand.type_name)
-        if isinstance(contains_value, dict):
-            return contains_value
-        cast = "::jsonb" if _is_json_type(operand.type_name) else ""
-        return f"{operand.sql} @> {_add_param(params, contains_value)}{cast}"
-    return err(f"Unknown row-read operator: {operator}", code=INVALID_OPERATOR)
-
-
-def _compile_order(
-    order_value: str | None,
-    column_meta: dict[str, str],
-    params: list[Any],
-) -> str | dict[str, Any]:
-    if not order_value:
-        return ""
-    parts: list[str] = []
-    for token in _split_top_level(order_value):
-        bits = token.rsplit(".", 1)
-        direction = "ASC"
-        field = token
-        if len(bits) == 2 and bits[1].lower() in {"asc", "desc"}:
-            field, direction = bits[0], bits[1].upper()
-        operand_or_error = _compile_operand(field, column_meta)
-        if isinstance(operand_or_error, dict):
-            return operand_or_error
-        operand = _bind_operand_params(operand_or_error, params)
-        parts.append(f"{operand.sql} {direction}")
-    return ", ".join(parts)
-
-
-def _compile_operand(raw: str, column_meta: dict[str, str]) -> _Operand | dict[str, Any]:
-    token = raw.strip()
-    m = _JSON_PATH_RE.fullmatch(token)
-    if not m:
-        return _unknown_column(token, column_meta)
-    base = m.group("base")
-    if base not in column_meta:
-        return _unknown_column(base, column_meta)
-    arrow = m.group("arrow")
-    cast = m.group("cast")
-    if cast and cast not in CAST_SQL:
-        return err(f"Invalid JSON cast {cast!r}.", code=INVALID_CAST, allowed_casts=sorted(CAST_SQL))
-    if not arrow:
-        if cast:
-            return err("Casts are only supported for JSON path operands.", code=INVALID_CAST)
-        ident = table_data_repo.safe_ident(base)
-        return _Operand(sql=ident, params=[], type_name=column_meta[base])
-    if not _is_json_type(column_meta[base]):
-        return err(f"Column {base!r} is not a JSON column.", code=UNDEFINED_COLUMN)
-    path = (m.group("path") or "").strip()
-    if not path:
-        return err(f"Invalid JSON path operand: {raw}", code=INVALID_FILTER)
-    sql_base = table_data_repo.safe_ident(base)
-    if arrow == "->>":
-        expr = f"{sql_base} ->> ${{param}}::text"
-        operand_params: list[Any] = [path]
-    else:
-        path_items = _parse_json_path_list(path)
-        if isinstance(path_items, dict):
-            return path_items
-        expr = f"{sql_base} #>> ${{param}}::text[]"
-        operand_params = [path_items]
-    type_name = cast or "text"
-    if cast:
-        expr = f"({expr})::{CAST_SQL[cast]}"
-    return _Operand(sql=expr, params=operand_params, type_name=type_name)
-
-
-def _bind_operand_params(operand: _Operand, params: list[Any]) -> _Operand:
-    sql = operand.sql
-    for value in operand.params:
-        sql = sql.replace("${param}", _add_param(params, value), 1)
-    return _Operand(sql=sql, params=[], type_name=operand.type_name)
-
-
-def _add_param(params: list[Any], value: Any) -> str:
-    params.append(value)
-    return f"${len(params)}"
-
-
-def _parse_page(query_params: Sequence[tuple[str, str]], range_header: str | None) -> _Page | dict[str, Any]:
-    limit = _parse_int(_last_value(query_params, "limit"), default=DEFAULT_LIMIT)
-    offset = _parse_int(_last_value(query_params, "offset"), default=0)
-    if isinstance(limit, dict):
-        return limit
-    if isinstance(offset, dict):
-        return offset
-    if range_header:
-        m = re.fullmatch(r"\s*(\d+)-(\d+)\s*", range_header)
-        if not m:
-            return err("Invalid Range header; expected N-M.", code=INVALID_ARGUMENT)
-        start, end = int(m.group(1)), int(m.group(2))
-        if end < start:
-            return err("Invalid Range header; end must be >= start.", code=INVALID_ARGUMENT)
-        offset = start
-        limit = end - start + 1
-    if limit < 0 or offset < 0:
-        return err("limit and offset must be non-negative.", code=INVALID_ARGUMENT)
-    return _Page(limit=min(limit, MAX_LIMIT), offset=offset)
-
-
-def _parse_int(value: str | None, *, default: int) -> int | dict[str, Any]:
-    if value in {None, ""}:
-        return default
-    try:
-        return int(str(value))
-    except ValueError:
-        return err(f"Expected integer, got {value!r}.", code=INVALID_ARGUMENT)
-
-
-def _prefer_count_exact(prefer_header: str | None) -> bool:
-    return bool(prefer_header and "count=exact" in prefer_header.lower())
-
-
-def _last_value(query_params: Sequence[tuple[str, str]], key: str) -> str | None:
-    values = [v for k, v in query_params if k == key]
-    return values[-1] if values else None
-
-
-def _split_operator(raw_value: str) -> tuple[str | None, str]:
-    if "." not in raw_value:
-        return None, raw_value
-    op, value = raw_value.split(".", 1)
-    return op.lower(), value
-
-
-def _split_bool_condition(raw: str) -> tuple[str | None, str | None]:
-    first = raw.find(".")
-    if first == -1:
-        return None, None
-    return raw[:first], raw[first + 1:]
-
-
-def _nested_group(raw: str) -> tuple[str, str] | None:
-    for name in ("or", "and"):
-        prefix = f"{name}("
-        if raw.startswith(prefix) and raw.endswith(")"):
-            return name, raw[len(name):]
-    return None
-
-
-def _split_top_level(value: str) -> list[str]:
-    parts: list[str] = []
-    buf: list[str] = []
-    depth = 0
-    brace_depth = 0
-    for ch in value:
-        if ch == "(":
-            depth += 1
-        elif ch == ")":
-            depth = max(0, depth - 1)
-        elif ch == "{":
-            brace_depth += 1
-        elif ch == "}":
-            brace_depth = max(0, brace_depth - 1)
-        if ch == "," and depth == 0 and brace_depth == 0:
-            parts.append("".join(buf))
-            buf = []
-        else:
-            buf.append(ch)
-    parts.append("".join(buf))
-    return parts
-
-
-def _parse_json_path_list(raw: str) -> list[str] | dict[str, Any]:
-    text = raw.strip()
-    if not (text.startswith("{") and text.endswith("}")):
-        return err("#>> JSON path must use {a,b} syntax.", code=INVALID_FILTER)
-    return [p.strip() for p in text[1:-1].split(",") if p.strip()]
-
-
-def _parse_in_values(raw: str, type_name: str) -> list[Any] | dict[str, Any]:
-    text = raw.strip()
-    if not (text.startswith("(") and text.endswith(")")):
-        return err("in operator expects parenthesized values, e.g. in.(a,b).", code=INVALID_FILTER)
-    out: list[Any] = []
-    for item in _split_top_level(text[1:-1]):
-        converted = _convert_value(item.strip(), type_name)
-        if isinstance(converted, dict):
-            return converted
-        out.append(converted)
-    return out
-
-
-def _parse_contains_value(raw: str, type_name: str) -> Any | dict[str, Any]:
-    if not _is_json_type(type_name):
-        return err("cs operator is only supported for JSON columns in this table API.", code=INVALID_OPERATOR)
-    text = raw.strip()
-    if text.startswith("{") and text.endswith("}") and ":" not in text:
-        return json.dumps([p.strip() for p in text[1:-1].split(",") if p.strip()])
-    try:
-        return json.dumps(json.loads(text))
-    except ValueError:
-        return err("cs on JSON columns expects JSON or {a,b} syntax.", code=INVALID_FILTER)
-
-
-def _convert_value(raw: str, type_name: str) -> Any | dict[str, Any]:
-    try:
-        if type_name in {"text", "json", "jsonb"}:
-            return raw
-        if type_name in {"number", "numeric"}:
-            return Decimal(raw)
-        if type_name == "int":
-            return int(raw)
-        if type_name == "float":
-            return float(raw)
-        if type_name in {"boolean", "bool"}:
-            lowered = raw.lower()
-            if lowered in {"true", "t", "1"}:
-                return True
-            if lowered in {"false", "f", "0"}:
-                return False
-            raise ValueError
-        if type_name == "date":
-            return date.fromisoformat(raw)
-        if type_name == "timestamp":
-            return datetime.fromisoformat(raw.replace("Z", "+00:00"))
-        if type_name == "uuid":
-            return uuid.UUID(raw)
-    except (ValueError, InvalidOperation):
-        return err(f"Could not convert {raw!r} to {type_name}.", code=INVALID_ARGUMENT)
-    return raw
-
-
-def _unknown_column(name: str, column_meta: dict[str, str]) -> dict[str, Any]:
-    available = sorted(column_meta)
-    return err(
-        f"Column {name!r} does not exist on this table.",
-        code=UNDEFINED_COLUMN,
-        hint=fuzzy_hint(name, available, label="columns"),
-        available_columns=available,
-    )
-
-
-def _shape_result(
-    result: dict[str, Any],
-    *,
-    vault_name: str,
-    table_name: str,
-    projections: Sequence[_Projection],
-    count_exact: bool,
-    offset: int,
-) -> tuple[dict[str, Any], str | None]:
-    rows = result.get("items", [])
-    total = len(rows)
-    if count_exact and rows:
-        total = int(rows[0].get("__akb_total") or 0)
-    visible_rows = [
-        row for row in rows
-        if not count_exact or row.get("__akb_present")
-    ]
-    shaped_items = [_shape_item(row, projections) for row in visible_rows]
-    columns = _output_columns(result.get("columns", []), projections)
-    body = {
-        "kind": "table_query",
-        "vault": vault_name,
-        "table": table_name,
-        "columns": columns,
-        "items": shaped_items,
-        "total": total if count_exact else len(shaped_items),
-    }
-    if not count_exact:
-        return body, None
-    if shaped_items:
-        return body, f"{offset}-{offset + len(shaped_items) - 1}/{total}"
-    return body, f"*/{total}"
-
-
-def _shape_item(row: dict[str, Any], projections: Sequence[_Projection]) -> dict[str, Any]:
-    if any(p.output_key == "*" for p in projections):
-        out = {
-            k: v for k, v in row.items()
-            if k not in {"__akb_total", "__akb_present"} and not k.startswith("__akb_col_")
-        }
-    else:
-        out = {}
-    for p in projections:
-        if p.output_key != "*":
-            out[p.output_key] = row.get(p.result_key)
-    return out
-
-
-def _output_columns(raw_columns: Iterable[str], projections: Sequence[_Projection]) -> list[str]:
-    if any(p.output_key == "*" for p in projections):
-        columns = [
-            c for c in raw_columns
-            if c not in {"__akb_total", "__akb_present"} and not c.startswith("__akb_col_")
-        ]
-    else:
-        columns = []
-    for p in projections:
-        if p.output_key != "*":
-            columns.append(p.output_key)
-    return columns
