+++ backend/app/services/table_service.py
@@ -14,15 +14,13 @@ names.
 
 from __future__ import annotations
 
-import hashlib
-import json
 import logging
 import re
 import uuid
 from collections.abc import AsyncIterator
 from contextlib import asynccontextmanager
 from datetime import datetime, timezone
-from typing import Any, cast
+from typing import Any
 
 import asyncpg
 
@@ -35,7 +33,7 @@ from app.services.index_service import (
     build_table_chunk, delete_table_chunks, write_source_chunks,
 )
 from app.services.role_sync import get_role_sync
-from app.services.uri_service import table_uri, vault_uri
+from app.services.uri_service import table_uri
 from app.services.user_sql_executor import (
     PermissionDeniedError,
     UniqueViolationError,
@@ -53,7 +51,7 @@ from app.util.errors import (
     UNIQUE_VIOLATION,
     VAULT_ARCHIVED,
 )
-from app.util.text import fuzzy_hint, to_nfc_any
+from app.util.text import fuzzy_hint
 
 # Re-exported helpers used by publication_service for the
 # `table_query` share path. Other modules import directly from
@@ -763,147 +761,6 @@ def _canonical_pg_type(logical_type: str | None) -> str:
     return " ".join(pg_type.lower().split())
 
 
-def _indexed_column_names(indexes: list[dict]) -> set[str]:
-    names: set[str] = set()
-    for idx in indexes:
-        for col in idx.get("columns", []):
-            if isinstance(col, dict):
-                name = col.get("name")
-            else:
-                name = col
-            if isinstance(name, str):
-                names.add(name)
-    return names
-
-
-def _unique_column_names(unique_keys: list[dict]) -> set[str]:
-    names: set[str] = set()
-    for key in unique_keys:
-        cols = key.get("columns", [])
-        if len(cols) == 1 and isinstance(cols[0], str):
-            names.add(cols[0])
-    return names
-
-
-def _build_table_schema(vault_name: str, table: dict, pg_types: dict[str, str]) -> dict:
-    columns = table_registry_repo.parse_columns(table.get("columns"))
-    unique_keys = table_registry_repo.parse_json_list(table.get("unique_keys"))
-    indexes = table_registry_repo.parse_json_list(table.get("indexes"))
-    indexed = _indexed_column_names(indexes)
-    unique = _unique_column_names(unique_keys)
-
-    column_items: list[dict] = []
-    missing_columns: list[str] = []
-    type_mismatches: list[dict] = []
-    registry_names = {
-        c["name"]
-        for c in columns
-        if isinstance(c, dict) and isinstance(c.get("name"), str)
-    }
-    system_columns = sorted(name for name in pg_types if name.lower() in _RESERVED)
-    extra_columns = sorted(
-        name
-        for name in pg_types
-        if name not in registry_names and name.lower() not in _RESERVED
-    )
-
-    for col in columns:
-        name = col["name"]
-        logical_type = col.get("type", "text")
-        expected_pg_type = _canonical_pg_type(logical_type)
-        actual_pg_type = pg_types.get(name)
-        drift: dict[str, object] = {
-            "missing": actual_pg_type is None,
-            "type_mismatch": False,
-        }
-        if actual_pg_type is None:
-            missing_columns.append(name)
-        else:
-            actual_norm = " ".join(actual_pg_type.lower().split())
-            drift["type_mismatch"] = actual_norm != expected_pg_type
-            if drift["type_mismatch"]:
-                type_mismatches.append({
-                    "column": name,
-                    "registry_type": logical_type,
-                    "expected_pg_type": expected_pg_type,
-                    "pg_type": actual_pg_type,
-                })
-
-        column_items.append({
-            "name": name,
-            "type": logical_type,
-            "required": bool(col.get("required", False)),
-            "default": col.get("default"),
-            "check": col.get("check"),
-            "enum": col.get("enum"),
-            "unique": bool(col.get("unique", False) or name in unique),
-            "index": bool(col.get("index", False) or name in indexed),
-            "references": col.get("references"),
-            "on_delete": col.get("on_delete"),
-            "pg_type": actual_pg_type,
-            "drift": drift,
-        })
-
-    drift_summary = {
-        "has_drift": bool(missing_columns or extra_columns or type_mismatches),
-        "missing_columns": missing_columns,
-        "extra_columns": extra_columns,
-        "type_mismatches": type_mismatches,
-    }
-
-    name = table["name"]
-    return {
-        "kind": "table_schema",
-        "uri": table_uri(vault_name, name, collection=table.get("collection")),
-        "vault": vault_name,
-        "collection": table.get("collection"),
-        "name": name,
-        "table": name,
-        "sql_name": table_data_repo.pg_short_name(name),
-        "description": table.get("description"),
-        "columns": column_items,
-        "unique_keys": unique_keys,
-        "indexes": indexes,
-        "pg_types": pg_types,
-        "system_columns": system_columns,
-        "drift": drift_summary,
-    }
-
-
-async def get_table_schema(vault_id: uuid.UUID, table_name: str) -> dict:
-    pool = await get_pool()
-    async with pool.acquire() as conn:
-        vault = await conn.fetchrow("SELECT name FROM vaults WHERE id = $1", vault_id)
-        if not vault:
-            raise NotFoundError("Vault", str(vault_id))
-        table = await table_registry_repo.find_by_name(conn, vault_id, table_name)
-        if not table:
-            raise NotFoundError("Table", table_name)
-        pg_name = table_data_repo.pg_table_name(vault["name"], table_name)
-        pg_types = await _fetch_column_meta(conn, {pg_name})
-        return _build_table_schema(vault["name"], table, pg_types)
-
-
-async def get_vault_schema(vault_id: uuid.UUID) -> dict:
-    pool = await get_pool()
-    async with pool.acquire() as conn:
-        vault = await conn.fetchrow("SELECT name FROM vaults WHERE id = $1", vault_id)
-        if not vault:
-            raise NotFoundError("Vault", str(vault_id))
-        rows = await table_registry_repo.list_for_vault(conn, vault_id)
-        tables = []
-        for table in rows:
-            pg_name = table_data_repo.pg_table_name(vault["name"], table["name"])
-            pg_types = await _fetch_column_meta(conn, {pg_name})
-            tables.append(_build_table_schema(vault["name"], table, pg_types))
-        return {
-            "kind": "vault_table_schema",
-            "vault": vault["name"],
-            "tables": tables,
-            "total": len(tables),
-        }
-
-
 async def drop_table(
     vault_id: uuid.UUID,
     table_name: str,
@@ -1578,325 +1435,6 @@ async def alter_table(
         "columns": columns,
         "unique_keys": existing_uks,
         "indexes": existing_idxs,
-    }
-
-
-_MIGRATION_COLUMN_KEYS = {
-    "name",
-    "type",
-    "required",
-    "default",
-    "check",
-    "enum",
-    "unique",
-    "index",
-    "references",
-    "on_delete",
-}
-
-_MIGRATION_ALTER_COLUMN_KEYS = {
-    "name",
-    "set_default",
-    "default",
-    "drop_default",
-    "set_check",
-    "check",
-    "drop_check",
-    "set_not_null",
-    "drop_not_null",
-    "set_enum",
-    "enum",
-    "rename_enum_values",
-    "enum_renames",
-}
-
-
-def table_migration_checksum(operations: list[dict]) -> str:
-    """Return the deterministic checksum for a structured migration op list."""
-    operations = _normalize_migration_operations(operations)
-    payload = json.dumps(
-        operations,
-        sort_keys=True,
-        separators=(",", ":"),
-        ensure_ascii=False,
-    )
-    return hashlib.sha256(payload.encode("utf-8")).hexdigest()
-
-
-def _normalize_migration_operations(operations: list[dict]) -> list[dict]:
-    normalized = to_nfc_any(operations)
-    _validate_migration_operations(normalized)
-    return cast(list[dict], normalized)
-
-
-def _validate_migration_operations(operations: list[dict]) -> None:
-    if not isinstance(operations, list) or not operations:
-        raise ValidationError("Migration body must be a non-empty JSON array of operations.")
-    for index, op in enumerate(operations):
-        if not isinstance(op, dict):
-            raise ValidationError(f"Migration operation #{index + 1} must be an object.")
-
-
-def _validate_idempotency_key(idempotency_key: str | None) -> str:
-    if not isinstance(idempotency_key, str) or not idempotency_key.strip():
-        raise ValidationError("Idempotency-Key header is required for table migrations.")
-    value = idempotency_key.strip()
-    try:
-        parsed = uuid.UUID(value)
-    except ValueError as e:
-        raise ValidationError("Idempotency-Key header must be a UUID.") from e
-    return str(parsed)
-
-
-def _migration_table_name(op: dict) -> str:
-    table_name = op.get("table") or op.get("table_name")
-    if not isinstance(table_name, str) or not table_name:
-        raise ValidationError("Migration operation requires a non-empty table/table_name.")
-    return table_name
-
-
-def _column_spec_from_migration_op(op: dict, *, keys: set[str], ctx: str) -> dict:
-    column = op.get("column")
-    if isinstance(column, dict):
-        spec = dict(column)
-    else:
-        spec = {key: op[key] for key in keys if key in op}
-        if isinstance(column, str) and "name" not in spec:
-            spec["name"] = column
-    if not spec:
-        raise ValidationError(f"{ctx} requires a column object or column fields.")
-    return spec
-
-
-def _object_spec_from_migration_op(op: dict, *, field: str, keys: set[str], ctx: str) -> dict:
-    value = op.get(field)
-    if isinstance(value, dict):
-        spec = dict(value)
-    else:
-        spec = {key: op[key] for key in keys if key in op}
-    if not spec:
-        raise ValidationError(f"{ctx} requires a {field} object or fields.")
-    return spec
-
-
-def _name_from_migration_op(op: dict, *fields: str, ctx: str) -> str:
-    for field in fields:
-        if field not in op:
-            continue
-        value = op[field]
-        if isinstance(value, dict):
-            value = value.get("name")
-        if isinstance(value, str) and value:
-            return value
-    raise ValidationError(f"{ctx} requires a non-empty name.")
-
-
-def _migration_op_to_alter_kwargs(op: dict) -> tuple[str, dict[str, Any]]:
-    raw_op = op.get("op")
-    if not isinstance(raw_op, str) or not raw_op:
-        raise ValidationError("Migration operation requires a non-empty op.")
-    op_name = raw_op.strip().lower().replace("-", "_")
-    table_name = _migration_table_name(op)
-
-    if op_name == "add_column":
-        return table_name, {
-            "add_columns": [
-                _column_spec_from_migration_op(
-                    op,
-                    keys=_MIGRATION_COLUMN_KEYS,
-                    ctx="add_column",
-                )
-            ],
-        }
-    if op_name == "alter_column":
-        return table_name, {
-            "alter_columns": [
-                _column_spec_from_migration_op(
-                    op,
-                    keys=_MIGRATION_ALTER_COLUMN_KEYS,
-                    ctx="alter_column",
-                )
-            ],
-        }
-    if op_name == "drop_column":
-        return table_name, {
-            "drop_columns": [
-                _name_from_migration_op(op, "name", "column", ctx="drop_column")
-            ],
-        }
-    if op_name == "rename_column":
-        old_name = _name_from_migration_op(
-            op, "from", "old_name", "from_name", "old", "column", ctx="rename_column.from"
-        )
-        new_name = _name_from_migration_op(
-            op, "to", "new_name", "to_name", "new", ctx="rename_column.to"
-        )
-        return table_name, {"rename_columns": {old_name: new_name}}
-    if op_name == "add_unique_key":
-        return table_name, {
-            "add_unique_keys": [
-                _object_spec_from_migration_op(
-                    op,
-                    field="unique_key",
-                    keys={"name", "columns"},
-                    ctx="add_unique_key",
-                )
-            ],
-        }
-    if op_name == "drop_unique_key":
-        return table_name, {
-            "drop_unique_keys": [
-                _name_from_migration_op(op, "name", "unique_key", ctx="drop_unique_key")
-            ],
-        }
-    if op_name == "add_index":
-        return table_name, {
-            "add_indexes": [
-                _object_spec_from_migration_op(
-                    op,
-                    field="index",
-                    keys={"name", "columns"},
-                    ctx="add_index",
-                )
-            ],
-        }
-    if op_name == "drop_index":
-        return table_name, {
-            "drop_indexes": [
-                _name_from_migration_op(op, "name", "index", ctx="drop_index")
-            ],
-        }
-    raise ValidationError(
-        "Unsupported migration op "
-        f"{raw_op!r}. Supported ops: add_column, alter_column, drop_column, "
-        "rename_column, add_unique_key, drop_unique_key, add_index, drop_index."
-    )
-
-
-async def apply_table_migration(
-    vault_id: uuid.UUID,
-    *,
-    actor_id: str,
-    idempotency_key: str | None,
-    operations: list[dict],
-) -> dict:
-    """Apply a structured table migration atomically and idempotently."""
-    key = _validate_idempotency_key(idempotency_key)
-    operations = _normalize_migration_operations(operations)
-    checksum = table_migration_checksum(operations)
-
-    pool = await get_pool()
-    changed_tables: dict[str, dict] = {}
-    async with pool.acquire() as conn:
-        async with conn.transaction():
-            vault = await conn.fetchrow("SELECT name FROM vaults WHERE id = $1", vault_id)
-            if not vault:
-                raise NotFoundError("Vault", str(vault_id))
-
-            await conn.execute(
-                "SELECT pg_advisory_xact_lock(hashtextextended($1, 0))",
-                f"{vault_id}:{key}",
-            )
-            existing = await conn.fetchrow(
-                """
-                SELECT id, applied_at, checksum
-                  FROM vault_migrations
-                 WHERE vault_id = $1 AND name = $2
-                 FOR UPDATE
-                """,
-                vault_id,
-                key,
-            )
-            if existing:
-                if existing["checksum"] != checksum:
-                    raise ConflictError(
-                        "Table migration idempotency key was already used with "
-                        "a different checksum."
-                    )
-                return {
-                    "kind": "table_migration",
-                    "vault": vault["name"],
-                    "idempotency_key": key,
-                    "checksum": checksum,
-                    "applied": False,
-                    "applied_at": existing["applied_at"].isoformat(),
-                    "operations": len(operations),
-                    "results": [],
-                }
-
-            results: list[dict[str, Any]] = []
-            for index, op in enumerate(operations, start=1):
-                table_name, kwargs = _migration_op_to_alter_kwargs(op)
-                result = await alter_table(
-                    vault_id,
-                    table_name,
-                    actor_id=actor_id,
-                    _conn=conn,
-                    _defer_index=True,
-                    **kwargs,
-                )
-                results.append({
-                    "index": index,
-                    "op": op["op"],
-                    "table": table_name,
-                    "result": result,
-                })
-                changed_tables[table_name] = result
-
-            row = await conn.fetchrow(
-                """
-                INSERT INTO vault_migrations (vault_id, name, checksum)
-                VALUES ($1, $2, $3)
-                RETURNING id, applied_at
-                """,
-                vault_id,
-                key,
-                checksum,
-            )
-            await emit_event(
-                conn,
-                "table.migration",
-                vault_id=vault_id,
-                resource_uri=vault_uri(vault["name"]),
-                actor_id=actor_id,
-                payload={
-                    "vault": vault["name"],
-                    "idempotency_key": key,
-                    "checksum": checksum,
-                    "operations": len(operations),
-                    "tables": sorted(changed_tables),
-                },
-            )
-
-    async with pool.acquire() as conn:
-        for table_name, result in changed_tables.items():
-            table = await table_registry_repo.find_by_name(conn, vault_id, table_name)
-            if not table:
-                continue
-            try:
-                await index_table_metadata(
-                    str(table["id"]),
-                    vault_id=vault_id,
-                    vault_name=result["vault"],
-                    name=table_name,
-                    description=table["description"] or "",
-                    columns=result["columns"],
-                    unique_keys=result.get("unique_keys"),
-                    indexes=result.get("indexes"),
-                )
-            except Exception as e:  # noqa: BLE001
-                logger.warning("table migration chunk reindex failed for %s: %s", table_name, e)
-
-    return {
-        "kind": "table_migration",
-        "id": str(row["id"]),
-        "vault": vault["name"],
-        "idempotency_key": key,
-        "checksum": checksum,
-        "applied": True,
-        "applied_at": row["applied_at"].isoformat(),
-        "operations": len(operations),
-        "results": results,
     }
