+++ backend/tests/test_table_migrations_unit.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import pytest
 
 from app.exceptions import ValidationError
-from app.services import table_service
+from app.services import table_migration_service
 
 
 def test_table_migration_checksum_is_stable_for_json_key_order() -> None:
@@ -28,7 +28,7 @@ def test_table_migration_checksum_is_stable_for_json_key_order() -> None:
         }
     ]
 
-    assert table_service.table_migration_checksum(ops_a) == table_service.table_migration_checksum(ops_b)
+    assert table_migration_service.table_migration_checksum(ops_a) == table_migration_service.table_migration_checksum(ops_b)
 
 
 def test_table_migration_operations_are_nfc_normalized_before_checksum() -> None:
@@ -51,15 +51,15 @@ def test_table_migration_operations_are_nfc_normalized_before_checksum() -> None
         }
     ]
 
-    assert table_service.table_migration_checksum(ops_nfd) == table_service.table_migration_checksum(ops_nfc)
-    normalized = table_service._normalize_migration_operations(ops_nfd)
-    table, kwargs = table_service._migration_op_to_alter_kwargs(normalized[0])
+    assert table_migration_service.table_migration_checksum(ops_nfd) == table_migration_service.table_migration_checksum(ops_nfc)
+    normalized = table_migration_service._normalize_migration_operations(ops_nfd)
+    table, kwargs = table_migration_service._migration_op_to_alter_kwargs(normalized[0])
     assert table == "café"
     assert kwargs["add_columns"][0]["default"] == "é"
 
 
 def test_table_migration_op_routes_to_alter_kwargs() -> None:
-    table, kwargs = table_service._migration_op_to_alter_kwargs({
+    table, kwargs = table_migration_service._migration_op_to_alter_kwargs({
         "op": "add_column",
         "table": "incidents",
         "name": "status",
@@ -82,7 +82,7 @@ def test_table_migration_op_routes_to_alter_kwargs() -> None:
         ]
     }
 
-    table, kwargs = table_service._migration_op_to_alter_kwargs({
+    table, kwargs = table_migration_service._migration_op_to_alter_kwargs({
         "op": "rename_column",
         "table_name": "incidents",
         "from": "status",
@@ -91,7 +91,7 @@ def test_table_migration_op_routes_to_alter_kwargs() -> None:
     assert table == "incidents"
     assert kwargs == {"rename_columns": {"status": "state"}}
 
-    table, kwargs = table_service._migration_op_to_alter_kwargs({
+    table, kwargs = table_migration_service._migration_op_to_alter_kwargs({
         "op": "alter_column",
         "table": "incidents",
         "column": {"name": "state", "set_default": "done", "drop_check": True},
@@ -102,15 +102,15 @@ def test_table_migration_op_routes_to_alter_kwargs() -> None:
 
 def test_table_migration_rejects_unknown_ops_and_bad_idempotency_key() -> None:
     with pytest.raises(ValidationError):
-        table_service._migration_op_to_alter_kwargs({"op": "sql", "table": "incidents"})
+        table_migration_service._migration_op_to_alter_kwargs({"op": "sql", "table": "incidents"})
 
     with pytest.raises(ValidationError):
-        table_service._validate_idempotency_key(None)
+        table_migration_service._validate_idempotency_key(None)
 
     with pytest.raises(ValidationError):
-        table_service._validate_idempotency_key("not-a-uuid")
+        table_migration_service._validate_idempotency_key("not-a-uuid")
 
     assert (
-        table_service._validate_idempotency_key("AAAAAAAA-AAAA-4AAA-8AAA-AAAAAAAAAAAA")
+        table_migration_service._validate_idempotency_key("AAAAAAAA-AAAA-4AAA-8AAA-AAAAAAAAAAAA")
         == "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa"
     )
